Upload lychee results to GH artifacts for excessively long reports (#386)
* Check report char count before adding to PR

Checks the char count of the lychee broken link report prior to
attaching to the PR. This is to help avoid a scenario where the report
hits the 65536 char max count of the GH API

* Replace hard coded max char count with variable

This allows us more flexibility to adjust the max length in the future
without doing a find-and-replace accross the workflow file

* Compare char count using fromJson function

This compares the char count to the action variable numerically (integer to
integer) rather than lexicographically. In this case, the report char count was
'1209122' which is _technically_ greater than 65536(the configured max), but
since the values were being compared lexicographically, and the string '1209122'
begins with a 1, this expression was evaluating to '1209122' >= '65536'(since
'1209122' begins with a 1).

* Reword report link comment

"Download full report" is a more actionable sentence than "See <link>
for full report"