Rename ptfe-releases to terraform-enterprise (#419)
* Rename PTFE Release Folder

Renames PTFE releases folder to terraform-enterprise

* Change PTFE Releases Product Config to TFE

* Redirect /ptfe-releases to /terraform-enterprise

* Update reference to ptfe-releases

Update code comments to make reference to terraform-enterprise, not
pfe-releases

* Add missing license header

* Update mermaid README diagram

Terraform is about to be migrated to UDR, so this diagram needed to be
updated to reflect that

* Refactor to remove references to ptfe-releases

Since #425 got merged into the base branch before this PR, additional
references to `/ptfe-releases` were re-introduced. These needed to be
removed in order for things to build and load correctly.